asteroids can be spawned at spawnpoints and velocity starts at direction to player. new direction_to() method of vector2

+++ src/entities/node.hpp
@@ -19,7 +19,7 @@ public:
 
     Vector2 position;
     Vector2 velocity;
-    float rotation_degrees;
+    double rotation_degrees;
     int width;
     int height;
